refactor(T-016/T-017): unify uniform maps and dev parameter registry
